move polyomino to a module
